pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isms-p-pii-toolkit"
version = "0.1.0"
description = "ISMS-P 관점의 개인정보 탐지 및 비식별화 학습용 툴킷"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "parkjinhan" }
]
dependencies = [
  "pycryptodome>=3.20.0",
  "fastapi>=0.115.0",
  "python-multipart>=0.0.9",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
isms-pii = "isms_pii_toolkit.cli:main"
isms-pii-api = "isms_pii_toolkit.api:run"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
isms_pii_toolkit = ["*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "--cov=isms_pii_toolkit",
  "--cov-report=term-missing",
  "--cov-fail-under=85",
]

[tool.coverage.run]
source = ["isms_pii_toolkit"]
branch = false

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]
show_missing = true
